src/test/kotlin/ru/romanow/FirstBadVersionTest.kt
package ru.romanow

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.Arguments.of
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource
import java.util.stream.Stream

class FirstBadVersionTest {

    @ArgumentsSource(ValueProvider::class)
    @ParameterizedTest(name = "#{index} – Expect broken version {0} in array {1}")
    fun firstBadVersion(n: Int, brokenVersion: Int) {
        val obj = FirstBadVersion(brokenVersion)
        assertThat(obj.firstBadVersion(n)).isEqualTo(brokenVersion)
    }

    internal class ValueProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext): Stream<Arguments> =
            Stream.of(of(5, 4), of(1, 1))
    }
}
